Fixed Activity File Names: Change from 'NewYorkTimes' to 'XingyunMain'
