+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,6 @@
         android:theme="@style/Theme.AndroidApplicationProject"
         tools:targetApi="31">
         <activity
-            android:name=".NewYorkTimes"
-            android:exported="false">
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
-        </activity>
-        <activity
             android:name=".XingyunMain"
             android:exported="false">
             <meta-data
